test: move all `substitute_vars` tests into `test_substitute_vars.py`
Right now there are some tests defined in `test_substitute_vars.py`
and some in `test_directive.py`. Move them all into a single file.
